# src/main/java/com/magazineluiza/favoritos/domain/user/LoginResponseDTO.java
package com.magazineluiza.favoritos.domain.user;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Setter
@Getter
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponseDTO {

	String token;
}