# swagger.json
openapi: 3.0.1
info:
  title: Api favorita Aiqfome
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
  /api/v1/clients/{id}:
    get:
      tags:
        - Clientes
      summary: Busca um cliente por ID
      description: Retorna os detalhes de um cliente específico usando seu ID.
      operationId: getClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
    put:
      tags:
        - Clientes
      summary: Atualiza um cliente existente
      description: Atualiza os dados de um cliente pelo seu ID. O e-mail, se alterado, deve ser único.
      operationId: updateClient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        '200':
          description: Cliente atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: 'Requisição inválida: Dados de atualização ausentes ou incorretos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: 'Conflito: E-mail já cadastrado para outro cliente'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
    delete:
      tags:
        - Clientes
      summary: Exclui um cliente
      description: Remove um cliente do sistema pelo seu ID.
      operationId: deleteClient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente excluído com sucesso (sem conteúdo na resposta)
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /auth/v1/register:
    post:
      summary: Registra um novo usuário
      description: Permite que um novo usuário crie uma conta no sistema fornecendo um login e senha. O login deve ser único
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDTO'
        required: true
      responses:
        '200':
          description: Registro bem-sucedido
          content:
            application/json: {}
        '400':
          description: 'Requisição inválida: Dados de registro inválidos ou ausentes'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: 'Conflito: Login já em uso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /auth/v1/login:
    post:
      summary: Autentica um usuário e retorna um token JWT
      description: Permite que um usuário existente faça login fornecendo suas credenciais (username/email e senha). Em caso de sucesso, um token de acesso JWT é retornado para uso em requisições protegidas.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationDTO'
        required: true
      responses:
        '200':
          description: Autenticação bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Requisicao não permitida, usuario ou senha estao incorretos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/clients:
    get:
      tags:
        - Clientes
      summary: Lista todos os clientes
      description: Retorna uma lista de todos os clientes cadastrados no sistema.
      operationId: getAllClients
      responses:
        '200':
          description: Lista de clientes retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientList'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails' 
    post:
      tags:
        - Clientes
      summary: Cria um novo cliente
      description: Registra um novo cliente no sistema com um e-mail único.
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        '201':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: 'Requisição inválida: Dados do cliente ausentes ou incorretos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: 'Conflito: E-mail já cadastrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/favorites/client/{clientId}:
    get:
      tags:
        - Produtos Favoritos
      summary: Lista produtos favoritos de um cliente
      description: Retorna todos os produtos que um cliente específico favoritou.
      operationId: getFavoritesByClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de produtos favoritos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteList'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
components:
  schemas:
    ClientList:
      type: array
      items:
          $ref: '#/components/schemas/Client' 
...
    FavoriteList:
      type: array
      items:
          $ref: '#/components/schemas/FavoriteProductResponseDTO' 
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        details:
          type: string
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        cellPhone:
          type: string
        streetAddress:
          type: string
        addressComplement:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
    RegisterDTO:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    AuthenticationDTO:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    LoginResponseDTO:
      type: object
      properties:
        token:
          type: string
...
    FavoriteProductResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titulo:
          type: string
        imagem:
          type: string
        preco:
          type: number
        review:
          type: number
        clientId:
          type: string
          format: uuid
        productId:
          type: integer
          format: int64
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT